sistana: move/simplify util module
